refactor: ButtonSample -> Button으로 변경
