+++ frontend/src/components/common/Button/Button.styled.ts
@@ -1,32 +1,62 @@
 import styled from 'styled-components';
+import type { ButtonSize, ButtonVariant } from './Button';
 
 interface CommonButtonProps {
-  $bgColor: string;
-  $fontColor: string;
-  $hoverColor: string;
+  $size: ButtonSize;
+  $variant: ButtonVariant;
 }
 
-export const CommonButton = styled.button<CommonButtonProps>`
-  background-color: var(${(props) => props.$bgColor});
-  color: var(${(props) => props.$fontColor});
+const buttonSize = (size: ButtonSize) => {
+  switch (size) {
+    case 'half':
+      return '27.1rem';
 
-  width: fit-content;
-  padding: 1.2rem 1.8rem;
+    case 'full':
+      return '100%';
+
+    default:
+      return 'fit-content';
+  }
+};
+
+const color = (variant: ButtonVariant) => {
+  switch (variant) {
+    case 'primary':
+      return `background-color: var(--primary-500);
+              color: var(--white-color);
+                &:hover {
+                  background-color: var(--primary-600);
+                }
+            `;
+
+    default:
+      return `background-color: var(--grey-200);
+              color: var(--black-color);
+                &:hover {
+                  background-color: var(--grey-300);
+                }
+            `;
+  }
+};
+
+export const CommonButton = styled.button<CommonButtonProps>`
+  ${(props) => color(props.$variant)}
+
+  width: ${(props) => buttonSize(props.$size)};
+  min-height: 4.2rem;
+  padding: 1rem 1.4rem;
   border-radius: 0.8rem;
 
   display: flex;
   gap: 0.3rem;
   justify-content: center;
   align-items: center;
+  transition: 0.2s;
 
   white-space: nowrap;
   font-size: 1.4rem;
   font-weight: 500;
   font-family: inherit;
-
-  &:hover {
-    background-color: var(${(props) => props.$hoverColor});
-  }
 
   &:disabled {
     cursor: default;
